Revert "moved issue related classes to de.uka.ipd.sdq.errorhandling.core"
This reverts commit 98e667295a378a576d91bd40604e198b154f33fc.
