+++ bundles/de.uka.ipd.sdq.errorhandling/src/de/uka/ipd/sdq/errorhandling/SeverityEnum.java
@@ -1,4 +1,4 @@
-package de.uka.ipd.sdq.errorhandling.core;
+package de.uka.ipd.sdq.errorhandling;
 
 public enum SeverityEnum {
     ERROR, WARNING
